Move TZE transaction builder to an independent module.
